DjangoPyServer/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
construct==2.5.3
dj-database-url==0.5.0
Django==3.0.6
django-admin-display==1.1.0
django-admin-tools==0.9.0
django-geojson==3.0.0
django-leaflet==0.26.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
future==0.18.2
haystack==0.42
idna==2.9
jsonfield==3.1.0
leaflet==0.0.3
mysql.connector==2.2.9
mysqlclient==1.4.6
ordered-set==4.0.1
packaging==20.4
pefile==2019.4.18
pyparsing==2.4.7
python-ptrace==0.9.5
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
